Modular isoch audio pipeline, Duet support, TX verification, and clock sync
Isochronous subsystem refactored into layered architecture:

DMA layer (new, hardware-only, no audio semantics):
- IsochRxDmaRing: generic IR descriptor engine, replaces monolithic RX context
- IsochTxDmaRing: generic IT engine with two-cursor refill and silence ring
- IsochTxDescriptorSlab: OHCI IT descriptor slab allocator
- IsochReceiveContext relocated to Isoch/Receive/ (parallel to Transmit/)

Audio pipeline layer (new, above DMA):
- IsochAudioRxPipeline: CIP parsing, AM824 decoding, shared queue pump,
  external sync correlation
- IsochAudioTxPipeline: AM824 encoding, SYT generation, sample timing

Audio driver (refactored and expanded):
- AudioClockEngine: 48kHz clock locking against OHCI cycle timer (PID control)
- AudioIOPath: zero-copy buffer handoff and sample time correlation
- AudioDriverConfig / AudioDriverConfigPolicy: device-specific port
  provisioning and format negotiation with quirk support
- AudioControlBuilder: DriverKit control nub registration for DAC/ADC/mixer
- AudioSharedMemoryBridge: SPSC queue bridge between pipeline and IOPath
- ASFWProtocolBooleanControl: DriverKit boolean control nub (phantom, polarity)

External sync (new):
- ExternalSyncBridge: OHCI cycle timer to wallclock correlation
- ExternalSyncDiscipline48k: PID frequency/phase lock for 48kHz streams
- RxBufferProfiles: configurable RX buffer sizing profiles

TX verification and recovery (new):
- IsochTxVerifier: watchdog-driven packet loss and corruption detection
- IsochTxRecoveryController: graceful underrun and corruption recovery
- TxVerifierDecode: TX capture and decode helpers

Device protocol: Apogee Duet
- Full Apogee Duet protocol: phantom power, polarity, DSP settings, stream-
  mode quirk, hardware-backed DriverKit controls
- DeviceProtocolFactory extended for Duet instantiation
- Oxford/Apogee stub foundations

Bug fixes:
- Fixed 7 critical bugs in Focusrite Saffire PRO 24 DSP DICE implementation
- Unified cycle-time clock recovery for TX and RX streams

Swift app:
- DuetControlView/ViewModel/Models: SwiftUI panel for Duet hardware controls
- DriverConnector+Duet: userspace connector extension for Duet commands
- Raw AV/C command tab for debugging
- Independent isochronous verbosity and telemetry logging controls

C++ modernization:
- IONew()/IOSafeDeleteNULL() throughout (DriverKit memory model)
- std::make_shared<> for shared ownership (IRM, CMP clients)
- const-correctness and init-statement scoping

Test coverage (+13 files):
- IsochRxDmaRingTests, IsochTxDescriptorSlabTests
- AudioIOPathTests, AudioDriverConfigPolicyTests
- ExternalSyncBridgeTests, ExternalSyncDiscipline48kTests
- SYTGeneratorTests, TxVerifierDecodeTests
- ApogeeBooleanControlMappingTests, DeviceProtocolFactoryTests
- DuetControlViewModelTests, DuetCodecTests, AVCUnitWireParsingTests
- AudioDriverKit mock header for host-side DriverKit testing

Removed: stale architecture .md files (ACK_REFACTOR, AUDIODRIVER, AVC,
TLABEL, ASYNC_ACK_DEBUGGING)

+++ ASFW/Models/DriverConnectorModels.swift
@@ -187,8 +187,8 @@ struct DriverConnectorAVCUnitInfo: Identifiable {
     let id = UUID()
     let guid: UInt64
     let nodeID: UInt16
-    let vendorID: UInt16
-    let modelID: UInt16
+    let vendorID: UInt32
+    let modelID: UInt32
     let subunits: [DriverConnectorAVCSubunitInfo]
     
     // Unit-level plug counts (from AVCUnitPlugInfoCommand)
